* Compiler warning
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@2196 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ mailbox.cpp
@@ -198,7 +198,7 @@ bool Mailbox::getReceiver(Item* item, std::string& name, uint32_t& dp)
 	{
 		for(IntegerVec::iterator it = disabledTowns.begin(); it != disabledTowns.end(); ++it)
 		{
-			if(town->getTownID() == (*it))
+			if(town->getTownID() == (uint32_t)(*it))
 				return false;
 		}
 	}
